[*] moved PomoSettingsModal methods to utils
